exercicio 39  reescrto codigo
